fix: 대분류 페이지 카테고리 그리드 컴포넌트 분리
