+++ app/[locale]/community/council/page.tsx
@@ -1,4 +1,5 @@
-import MajorCategoryPageLayout from '@/components/layout/pageLayout/MajorCategoryPageLayout';
+import Header from '@/components/layout/header/Header';
+import CategoryGrid from '@/components/layout/pageLayout/CategoryGrid';
 import PageTitle from '@/components/layout/pageLayout/PageTitle';
 import { council } from '@/constants/segmentNode';
 import { getMetadata } from '@/utils/metadata';
@@ -13,16 +14,15 @@ export async function generateMetadata(props: { params: Promise<{ locale: string
 
 export default async function CouncilPage() {
   return (
-    <MajorCategoryPageLayout
-      theme="light"
-      titleComponent={
-        <PageTitle
-          title={council.name}
-          currentPage={council}
-          titleType="big"
-          margin="mb-6 sm:mb-11"
-        />
-      }
-    />
+    <div className="bg-neutral-850">
+      <Header />
+      <PageTitle
+        title={council.name}
+        currentPage={council}
+        titleType="big"
+        margin="mb-6 sm:mb-11"
+      />
+      <CategoryGrid currentPage={council} theme="light" />
+    </div>
   );
 }
